edge features to long when building graph

+++ src/data/graph_builder.py
@@ -159,20 +159,20 @@ def build_hetero_graph(edges: pd.DataFrame) -> Tuple[HeteroData, Dict[str, Dict[
             # Event-based: use edge_weight
             hetero_data[edge_type_key].edge_attr = torch.tensor(
                 group['edge_weight'].values,
-                dtype=torch.float
+                dtype=torch.long
             ).unsqueeze(-1)
         elif has_score:
             # Snapshot-based: use score
             hetero_data[edge_type_key].edge_attr = torch.tensor(
                 group['score'].values,
-                dtype=torch.float
+                dtype=torch.long
             ).unsqueeze(-1)
         
         # Add temporal attribute
         if has_edge_time:
             hetero_data[edge_type_key].edge_time = torch.tensor(
                 group['edge_time'].values,
-                dtype=torch.float
+                dtype=torch.long  # Use long for temporal sampling compatibility
             )
         
         num_edges = edge_index.size(1)
